+++ CHANGELOG.md
@@ -12,7 +12,7 @@ labeled as 2.7.1. Subsequent releases will follow
   *
 
 ### Fixed
-  *
+  * Fix validation of address checksum and prefix when encoding and decoding
   *
 
 ### Deprecated